[NFC] Move ModuleUtils::collectHeapTypes to a .cpp file
In preparation for the refactoring in #4455.
